Move files to standard file locations for fuzzing infra
