--- fuzz/compile-and-execute/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "fuzz-compile-and-execute"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-libfuzzer-sys = "0.4.0"
-rand = "0.8.5"
-rand_chacha = "0.3.1"
-clvmr = "0.1.21"
-clvm_tools_rs = { path = "../..", features = ["fuzzer"] }
